Foi adicionado o arquivo de css ao app.js
