+++ package.json
@@ -1,6 +1,7 @@
 {
   "scripts": {
     "dev": "next dev"
+    
   },
   "dependencies": {
     "next": "^12.0.7",
